requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.18.0
pandas>=1.0.0
PyYAML>=5.3.1
pydantic>=1.8.0
tqdm>=4.0.0
Pillow>=7.0.0